Change numbers in file names to integers

+++ p1_gnxsims/parallel/parallel_LGS.py
@@ -513,8 +513,8 @@ def run_sims(sim_list):
     print(params)
 
     # make our params dict into a proper Geonomics ParamsDict object
-    mod_name = "10k_K" + str(K) + "_phi" + str(int(phi * 100)) + "_m" + str(
-        int(m * 100)) + "_seed" + str(seed) + "_H" + str(int(H * 100)) + "_r" + str(int(r * 100))
+    mod_name = "10k_K" + str(int(K)) + "_phi" + str(int(phi * 100)) + "_m" + str(
+        int(m * 100)) + "_seed" + str(int(seed)) + "_H" + str(int(H * 100)) + "_r" + str(int(r * 100))
     print(mod_name)
     params = gnx.make_params_dict(params, mod_name)
     # then use it to make a model
